2. Telco Customer Churn/frontend/requirements.txt
requests
streamlit
typing_extensions
pils
pickle
pandas
json